+++ packages/typespec-rust/test/codegen.test.ts
@@ -51,9 +51,12 @@ describe('typespec-rust: codegen', () => {
 
   describe('helpers', () => {
     it('annotationDerive', () => {
-      strictEqual(helpers.annotationDerive(), '#[derive(Clone, Deserialize, SafeDebug, Serialize)]\n');
-      strictEqual(helpers.annotationDerive('Copy'), '#[derive(Clone, Copy, Deserialize, SafeDebug, Serialize)]\n');
-      strictEqual(helpers.annotationDerive('', 'Copy'), '#[derive(Clone, Copy, Deserialize, SafeDebug, Serialize)]\n');
+      strictEqual(helpers.annotationDerive(true), '#[derive(Clone, Deserialize, SafeDebug, Serialize)]\n');
+      strictEqual(helpers.annotationDerive(true, 'Copy'), '#[derive(Clone, Copy, Deserialize, SafeDebug, Serialize)]\n');
+      strictEqual(helpers.annotationDerive(true, '', 'Copy'), '#[derive(Clone, Copy, Deserialize, SafeDebug, Serialize)]\n');
+      strictEqual(helpers.annotationDerive(false), '#[derive(Clone, SafeDebug)]\n');
+      strictEqual(helpers.annotationDerive(false, 'Copy'), '#[derive(Clone, Copy, SafeDebug)]\n');
+      strictEqual(helpers.annotationDerive(false, '', 'Copy'), '#[derive(Clone, Copy, SafeDebug)]\n');
     });
 
     it('emitVisibility', () => {